requirements.txt
asyncio
beautifulsoup4
dotenv
python-telegram-bot
requests
schedule
spacy
pupputeer